+++ src/chanjo2/models/pydantic_models.py
@@ -6,9 +6,10 @@ from typing import Any, List, Optional
 import validators
 from pydantic import BaseModel, validator
 
-WRONG_COVERAGE_FILE_MSG = (
+WRONG_COVERAGE_FILE_MSG: str = (
     "Coverage_file_path must be either an existing local file path or a URL"
 )
+WRONG_BED_FILE_MSG: str = "Provided intervals files is not a valid BED file"
 
 
 class Builds(str, Enum):